Step1

+++ src/main/java/nextstep/courses/domain/CourseRepository.java
@@ -1,7 +1,8 @@
 package nextstep.courses.domain;
 
 public interface CourseRepository {
-    int save(Course course);
+
+    long save(Course course);
 
     Course findById(Long id);
 }
